refactor(templates): extract language tables from manifest god-object
Move the babel/BCP-47 alias tables and _map_babel_language/_map_bcp47_language
out of the 1039-line manifest.py into a focused core.templates.languages
module (manifest -> ~930 lines). runtime.py now imports the alias table from
languages instead of reaching into manifest, improving layering.

Pure relocation; all 849 tests green, ruff clean.
